table changed

+++ database/migrations/2020_01_08_155309_create_likes2_table.php
@@ -15,8 +15,8 @@ class CreateLikes2Table extends Migration
     {
         Schema::create('likes2', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id')->unsigned()->index();
-            $table->integer('article_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
+            $table->bigInteger('article_id')->unsigned()->index();
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
